refactor(gemini): add gemini module and service
