+++ apps/api/src/modules/gemini/gemini.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+
+import { GeminiService } from './gemini.service';
+
+@Module({
+  providers: [GeminiService],
+  exports: [GeminiService],
+})
+export class GeminiModule {}
