+++ apps/api/src/modules/onboarding/onboarding.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 
+import { GeminiModule } from '../gemini/gemini.module';
 import { OnboardingController } from './onboarding.controller';
 import { OnboardingService } from './onboarding.service';
 
 @Module({
+  imports: [GeminiModule],
   controllers: [OnboardingController],
   providers: [OnboardingService],
 })
